Tests/LibWeb/Text/expected/URL/url-search-params-sort-utf16-code-units.txt

%EF%BF%BC=&%EF%BF%BD=x&%EF%BF%BD=a
'\ufffc' => ''
'\ufffd' => '\u0078'
'\ufffd' => '\u0061'
